src/backend/requirements.txt
python-dotenv
fastapi
uvicorn
qdrant-client
pypdf
langchain-openai
langchain-qdrant
langchain-community
langchain-openai